refactor: restructure project into modular pattern
