+++ .gitignore
@@ -1,0 +1,37 @@
+/obj
+/dist
+*.exe
+*.o
+*.a
+*.lib
+*.dll
+*.so
+*.dylib
+
+/webserver
+/webserver.exe
+
+.vscode/
+.idea/
+*.sln
+*.vcxproj*
+*.vcxitems
+*.sdf
+*.opensdf
+*.suo
+*.user
+*.filters
+  
+*.pdb
+*.idb
+*.ilk
+*.exp
+*.db
+
+.DS_Store
+.DS_Store?
+._*
+.Spotlight-V100
+.Trashes
+ehthumbs.db
+Thumbs.db
